refactor [server]: move appModule to modules folder
